fix(workspace): share data-source state via context and add screenshots
The previous useDataSource hook used per-component useState, so
selecting a data source from the workspace did not propagate to the
new breadcrumb header. Hoist into a DataSourceProvider so AppShell and
QueryWorkspace stay in sync. Also add reference screenshots of the
new pages and dialog.
